Adds additional negative fulfillment tests
