+++ tests/unit/FulfillmentTest.php
@@ -2,17 +2,32 @@
 
 namespace Nikolag\Square\Tests\Unit;
 
+use Nikolag\Square\Exceptions\InvalidSquareOrderException;
+use Nikolag\Square\Facades\Square;
 use Nikolag\Square\Models\Fulfillment;
 use Nikolag\Square\Models\DeliveryDetails;
 use Nikolag\Square\Models\PickupDetails;
+use Nikolag\Square\Models\Product;
 use Nikolag\Square\Models\ShipmentDetails;
 use Nikolag\Square\Tests\Models\Order;
 use Nikolag\Square\Tests\TestCase;
+use Nikolag\Square\Tests\TestDataHolder;
 use Nikolag\Square\Utils\Constants;
 use Throwable;
 
 class FulfillmentTest extends TestCase
 {
+    private TestDataHolder $data;
+
+    /**
+     * @return void
+     */
+    public function setUp(): void
+    {
+        parent::setUp();
+        $this->data = TestDataHolder::make();
+    }
+
     /**
      * Make sure fulfillments cannot be created without fulfillment details.
      *
@@ -139,5 +154,47 @@ class FulfillmentTest extends TestCase
         $this->expectExceptionMessageMatches('/Integrity constraint violation/');
 
         factory(Fulfillment::class)->states(Constants::FULFILLMENT_TYPE_DELIVERY)->create();
+    }
+
+    /**
+     * Pickup creation without order, testing exception case.
+     *
+     * @return void
+     */
+    public function test_square_order_fulfillment_with_no_order(): void
+    {
+        $fulfillmentDetails = [
+            'delivery_details' => $this->data->fulfillmentWithDeliveryDetails,
+            'pickup_details'   => $this->data->fulfillmentWithPickupDetails,
+            'shipment_details' => $this->data->fulfillmentWithShipmentDetails,
+        ];
+        foreach ($fulfillmentDetails as $fulfillmentType => $fulfillment) {
+            // Retrieve the fulfillment with Shipment Details
+            $this->expectException(InvalidSquareOrderException::class);
+            $this->expectExceptionMessage('Fulfillment cannot be set without an order.');
+            $this->expectExceptionCode(500);
+
+            Square::setFulfillment($fulfillment);
+        }
+    }
+
+    /**
+     * Pickup creation without order, testing exception case.
+     *
+     * @return void
+     */
+    public function test_square_order_fulfillment_with_multiple_fulfillments(): void
+    {
+
+        $this->expectException(InvalidSquareOrderException::class);
+        $this->expectExceptionMessage('This order already has a fulfillment.');
+        $this->expectExceptionCode(500);
+
+        $product = factory(Product::class)->create();
+        Square::setOrder($this->data->order, env('SQUARE_LOCATION'))
+            ->addProduct($this->data->product, 1)
+            ->addProduct($product, 2)
+            ->setFulfillment($this->data->fulfillmentWithPickupDetails)
+            ->setFulfillment($this->data->fulfillmentWithPickupDetails);
     }
 }
